Renamed the os2borgerpc_client.os2borgerpc_config module to just "config"
I've written Java before; I like having to write the name of a module three
or four times as much as the next person. But we can do without that here
