+++ os2borgerpc_client/bin/get_os2borgerpc_config
@@ -4,7 +4,7 @@ from __future__ import print_function
 
 import sys
 
-from os2borgerpc_client.os2borgerpc_config import get_config
+from os2borgerpc_client.config import get_config
 
 DEBUG = True  # TODO: Get from command line option
