+++ os2borgerpc_client/bin/os2borgerpc_register_in_admin
@@ -5,7 +5,7 @@ from __future__ import print_function
 import sys
 import urlparse
 
-from os2borgerpc_client.os2borgerpc_config import OS2borgerPCConfig
+from os2borgerpc_client.config import OS2borgerPCConfig
 from os2borgerpc_client.admin_client import OS2borgerPCAdmin
